Fixed forward establish with flag. Improved logic.
